+++ atomicshop/wrappers/loggingw/checks.py
@@ -1,0 +1,20 @@
+import logging
+
+
+def has_handlers(logger: logging.Logger) -> bool:
+    """
+    Function to check if the logger has handlers.
+    :param logger: Logger to check
+    :return: True if logger has handlers, False otherwise
+    """
+
+    # Omitted the usage of "hasHandlers()" method, since sometimes returned "True" even when there were no handlers
+    # Didn't research the issue much, just used the "len(logger.handlers)" to check how many handlers there are
+    # in the logger.
+    # if not logging.getLogger(function_module_name).hasHandlers():
+    # if len(logging.getLogger(function_module_name).handlers) == 0:
+
+    if len(logger.handlers) == 0:
+        return False
+    else:
+        return True