+++ atomicshop/wrappers/loggingw/handlers.py
@@ -1,0 +1,133 @@
+import logging
+from logging.handlers import TimedRotatingFileHandler, QueueListener, QueueHandler
+import re
+
+
+def get_stream_handler() -> logging.StreamHandler:
+    """
+    Function to get a StreamHandler.
+    This handler that will output messages to the console.
+
+    :return: StreamHandler.
+    """
+
+    return logging.StreamHandler()
+
+
+def get_timed_rotating_file_handler(
+        log_file_path: str, when: str = "midnight", interval: int = 1, delay: bool = False
+) -> logging.handlers.TimedRotatingFileHandler:
+    """
+    Function to get a TimedRotatingFileHandler.
+    This handler will output messages to a file, rotating the log file at certain timed intervals.
+
+    :param log_file_path: Path to the log file.
+    :param when: When to rotate the log file. Possible values:
+        "S" - Seconds
+        "M" - Minutes
+        "H" - Hours
+        "D" - Days
+        "midnight" - Roll over at midnight
+    :param interval: Interval to rotate the log file.
+    :param delay: bool, If set to True, the log file will be created only if there's something to write.
+    :return: TimedRotatingFileHandler.
+    """
+
+    return TimedRotatingFileHandler(filename=log_file_path, when=when, interval=interval, delay=delay)
+
+
+def start_queue_listener_for_file_handler(
+        file_handler: logging.FileHandler, queue_object) -> logging.handlers.QueueListener:
+    """
+    Function to get a QueueListener for the FileHandler.
+    This handler get the messages from the FileHandler and put them in the Queue.
+
+    :param file_handler: FileHandler to get the messages from.
+    :param queue_object: Queue object to put the messages in.
+    :return: QueueListener.
+    """
+
+    # Create the QueueListener based on TimedRotatingFileHandler
+    queue_listener = QueueListener(queue_object, file_handler)
+    # Start the QueueListener. Each logger will have its own instance of the Queue
+    queue_listener.start()
+
+    return queue_listener
+
+
+def get_queue_handler(queue_object) -> logging.handlers.QueueHandler:
+    """
+    Function to get a QueueHandler.
+    This handler gets the messages from the Queue and writes them to file of the FileHandler that was set for
+    QueueListener.
+
+    :param queue_object: Queue object to get the messages from.
+    :return: QueueHandler.
+    """
+
+    return QueueHandler(queue_object)
+
+
+def set_formatter(handler: logging.Handler, logging_formatter: logging.Formatter):
+    """
+    Function to set the formatter for the handler.
+    :param handler: Handler to set the formatter to.
+    :param logging_formatter: logging Formatter to set to the handler.
+    """
+
+    handler.setFormatter(logging_formatter)
+
+
+def set_handler_name(handler: logging.Handler, handler_name: str):
+    """
+    Function to set the handler name.
+    :param handler: Handler to set the name to.
+    :param handler_name: Name to set to the handler.
+    """
+
+    handler.set_name(handler_name)
+
+
+def get_handler_name(handler: logging.Handler) -> str:
+    """
+    Function to get the handler name.
+    :param handler: Handler to get the name from.
+    :return: Handler name.
+    """
+
+    return handler.get_name()
+
+
+def change_rotated_filename(file_handler: logging.Handler, file_extension: str):
+    # Changing the way TimedRotatingFileHandler managing the rotating filename
+    # Default file suffix is only "Year_Month_Day" with addition of the dot (".") character to the
+    # "file name + extension" that you provide it. Example: log file name:
+    # test.log
+    # After file is rotated at midnight, by default the old filename will be:
+    # test.log.2021_12_24
+    # And the log file of 25th, now will be "test.log".
+    # So, Changing the file suffix to include the extension to the suffix, so it will be:
+    # test.log.2021_12_24.log
+    # file_handler.suffix = logfile_suffix
+    # file_handler.suffix = "_%Y_%m_%d.txt"
+    # This step will remove the created ".log." above before the suffix and the filename will look like:
+    # test.2021_12_24.log
+    # file_handler.namer = lambda name: name.replace(log_file_extension + ".", "") + log_file_extension
+    # file_handler.namer = lambda name: name.replace(".txt.", "") + log_file_extension
+    # This will recompile the string to tell the handler the length of the suffix parts
+    # file_handler.extMatch = re.compile(r"^\d{4}_\d{2}_\d{2}" + re.escape(log_file_extension) + r"$")
+    # file_handler.extMatch = re.compile(r"^\d{4}_\d{2}_\d{2}.txt$")
+
+    # Set variables that are responsible for setting TimedRotatingFileHandler filename on rotation.
+    # Log files time format, need only date
+    format_date_log_filename: str = "%Y_%m_%d"
+    # Log file suffix.
+    logfile_suffix: str = "_" + format_date_log_filename + file_extension
+    # Regex object to match the TimedRotatingFileHandler file name suffix.
+    # "re.escape" is used to "escape" strings in regex and use them as is.
+    logfile_regex_suffix = re.compile(r"^\d{4}_\d{2}_\d{2}" + re.escape(file_extension) + r"$")
+
+    # Changing the setting that we set above
+    file_handler.suffix = logfile_suffix
+    file_handler.namer = lambda name: name.replace(file_extension + ".", "") + file_extension
+    file_handler.extMatch = logfile_regex_suffix